+++ server/src/main/java/org/opensearch/action/admin/cluster/node/info/NodesInfoRequest.java
@@ -53,7 +53,7 @@ import java.util.stream.Collectors;
 @PublicApi(since = "1.0.0")
 public class NodesInfoRequest extends BaseNodesRequest<NodesInfoRequest> {
 
-    private Set<String> requestedMetrics = Metric.allMetrics();
+    private Set<String> requestedMetrics = Metric.defaultMetrics();
 
     /**
      * Create a new NodeInfoRequest from a {@link StreamInput} object.
@@ -73,7 +73,7 @@ public class NodesInfoRequest extends BaseNodesRequest<NodesInfoRequest> {
      */
     public NodesInfoRequest(String... nodesIds) {
         super(nodesIds);
-        all();
+        defaultMetrics();
     }
 
     /**
@@ -89,6 +89,15 @@ public class NodesInfoRequest extends BaseNodesRequest<NodesInfoRequest> {
      */
     public NodesInfoRequest all() {
         requestedMetrics.addAll(Metric.allMetrics());
+        return this;
+    }
+
+    /**
+     * Sets to return data for default metrics.
+     * See {@link Metric#defaultMetrics()}.
+     */
+    public NodesInfoRequest defaultMetrics() {
+        requestedMetrics.addAll(Metric.defaultMetrics());
         return this;
     }
 
@@ -156,7 +165,7 @@ public class NodesInfoRequest extends BaseNodesRequest<NodesInfoRequest> {
 
     /**
      * An enumeration of the "core" sections of metrics that may be requested
-     * from the nodes information endpoint. Eventually this list list will be
+     * from the nodes information endpoint. Eventually this list will be
      * pluggable.
      */
     public enum Metric {
@@ -171,7 +180,8 @@ public class NodesInfoRequest extends BaseNodesRequest<NodesInfoRequest> {
         INGEST("ingest"),
         AGGREGATIONS("aggregations"),
         INDICES("indices"),
-        SEARCH_PIPELINES("search_pipelines");
+        SEARCH_PIPELINES("search_pipelines"),
+        ANALYSIS_COMPONENTS("analysis_components");
 
         private String metricName;
 
@@ -187,8 +197,27 @@ public class NodesInfoRequest extends BaseNodesRequest<NodesInfoRequest> {
             return metricNames.contains(this.metricName());
         }
 
+        /**
+         * Return all available metrics.
+         */
         public static Set<String> allMetrics() {
             return Arrays.stream(values()).map(Metric::metricName).collect(Collectors.toSet());
+        }
+
+        /**
+         * Return "the default" set of metrics.
+         * Similar to {@link #allMetrics()} except {@link Metric#SEARCH_PIPELINES} and
+         * {@link Metric#ANALYSIS_COMPONENTS} metrics are not included.
+         * <br>
+         * The motivation to define the default set of metrics was to keep the default response
+         * size at bay. Metrics that are NOT included in the default set were typically introduced later
+         * and are considered to contain specific type of information that is not usually useful unless you
+         * know that you really need it.
+         */
+        public static Set<String> defaultMetrics() {
+            return allMetrics().stream()
+                .filter(metric -> !(metric.equals(ANALYSIS_COMPONENTS.metricName()) || metric.equals(SEARCH_PIPELINES.metricName())))
+                .collect(Collectors.toSet());
         }
     }
 }