+++ server/src/main/java/org/opensearch/node/Node.java
@@ -1165,7 +1165,8 @@ public class Node implements Closeable {
                 resourceUsageCollectorService,
                 segmentReplicationStatsTracker,
                 repositoryService,
-                admissionControlService
+                admissionControlService,
+                analysisModule.getAnalysisRegistry()
             );
 
             final SearchService searchService = newSearchService(