+++ server/src/test/java/org/opensearch/action/admin/cluster/stats/ClusterStatsNodesTests.java
@@ -181,6 +181,7 @@ public class ClusterStatsNodesTests extends OpenSearchTestCase {
             null,
             null,
             null,
+            null,
             null
         );
     }